feat: Enhance RPC client and server functionality with diagnostics and event notifications
- Improved error handling and logging in rpc_client.c for better diagnostics.
- Updated task_rpc.c to use a larger data type for consecutive RPC failures.
- Added configuration options in config.py for simulating processing delays.
- Implemented global score reconstruction in main.py to persist state across server restarts.
- Enhanced dashboard websocket communication to support typed events and payloads.
- Introduced asyncio task count display in the dashboard for better monitoring.
- Created demo_concurrency.sh to demonstrate concurrent RPC requests.
- Refactored GameManager and NodeRegistry to support event notifications for dashboard updates.
- Modified database queries to ensure consistent ordering of nodes.
- Added processing delay simulation in RPC server to test system behavior under load.

+++ firmware/tasks/task_rpc.c
@@ -175,7 +175,7 @@ wifi_connected:
   bool registered = false;
   TickType_t last_scores_refresh = xTaskGetTickCount();
   TickType_t last_register_attempt = 0;
-  uint8_t consecutive_rpc_failures = 0;
+  uint16_t consecutive_rpc_failures = 0;
   uint8_t consecutive_register_failures = 0;
   uint64_t last_heartbeat_us =
       0; /* Timestamp do último heartbeat bem-sucedido */
